+++ CHANGELOG.md
@@ -68,7 +68,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 
 ### New
 
-- TODO ([#XXX](https://github.com/kedacore/keda/issues/XXX))
+- **General**: Introduce AWS Systems Manager Parameter Store authentication provider ([#6311](https://github.com/kedacore/keda/issues/6311))
 
 #### Experimental